use termcolor for colorization
